fix: adjust number type to u16

+++ src/utils.rs
@@ -1,3 +1,3 @@
-pub fn normalize_bet_numbers(number: usize) -> String {
+pub fn normalize_bet_numbers(number: u16) -> String {
     String::from(format!("{:0zbefore$}", number, zbefore=4))       
 }
